Redis、レートリミット処理を変更
- 自前のリミットから、actix-limitationを使うように変更
- レートリミット機能にログを追加
- ArcでRedis接続

+++ backend/src/config/rate_limit.rs
@@ -4,7 +4,7 @@ use std::time::Duration;
 #[derive(Clone)]
 pub struct RateLimitConfig {
     pub duration: Duration,
-    pub max_requests: u32,
+    pub max_requests: u64,
 }
 
 impl RateLimitConfig {
